remove torch
